requirements.txt
gradio==5.7.1
groq==0.13.0
python-dotenv==1.0.1
